Update the iOS app as well
